postman

+++ src/utils/cloudinary.js
@@ -1,17 +1,23 @@
-import { v2 as cloudinary } from 'cloudinary'
-import fs from "fs";            //file system
+import {v2 as cloudinary} from "cloudinary"
+import fs from "fs"
+
+console.log(process.env.CLOUDINARY_CLOUD_NAME);
+console.log(process.env.CLOUDINARY_API_KEY);
+console.log(process.env.CLOUDINARY_API_SECRET);
 
 cloudinary.config({ 
-        cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
-        api_key:  process.env.CLOUDINARY_API_KEY, 
-        api_secret:  process.env.CLOUDINARY_API_SECRET
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
+  api_key: process.env.CLOUDINARY_API_KEY, 
+  api_secret: process.env.CLOUDINARY_API_SECRET 
 });
 
 const uploadOnCloudinary = async (localFilePath) => {
     try {
         if (!localFilePath) return null
         //upload the file on cloudinary
-        const response = await cloudinary.uploader.upload(localFilePath, {
+        const fixedPath = localFilePath.replace(/\\/g, "/");
+
+        const response = await cloudinary.uploader.upload(fixedPath, {
             resource_type: "auto"
         })
         // file has been uploaded successfull
@@ -20,9 +26,17 @@ const uploadOnCloudinary = async (localFilePath) => {
         return response;
 
     } catch (error) {
-        fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
-        return null;
+    console.log("CLOUDINARY ERROR:", error); // ⭐ ADD THIS
+    
+    if (fs.existsSync(localFilePath)) {
+        fs.unlinkSync(localFilePath);
     }
+
+    throw error; // ⭐ VERY IMPORTANT
 }
+
+}
+
+
 
 export {uploadOnCloudinary}
